CFN fixes for CI-CD pipeline
Updated the master-pipeline.yaml file to pass the BootstrapPrefix to the Pipeline CFN template as that value is used by the RestAPI resource created in the Pipeline template.  Also commented out the !If logic for the Lexv2 bot check in the Pipeline resource section as that did not logically need to be in place if the Pipeline CFN permitted a zero length BotName parameter.
Updated the pipeline.yaml file so that the BotName and BotAlias parameter definitions match what is established and working in the master-pipeline.yaml file.