Release 0.21.0 (#626)
* Vue 3 Migration (#620)

This PR is to upgrade the existing solution and relevant dependencies to Vue 3. Because we are also migrating from Vuetify 1.5 -> 3 in the process, skipping over Vuetify 2 entirely, every component required some clean-up. The goal was to get parity in UI from the previous release, which included some new CSS rules to maintain the look at feel of the previous UI.

These changes are likely to result in breaking changes to custom CSS modules, though again we have tried to maintain as much parity as possible its not feasible to test every CSS iteration end users might have implemented. Please use caution when updating to this version of the UI.

---------

Co-authored-by: Austin Johnson <austintj@amazon.com>, @michaelin-96